Add abstract class for task view holder
